rag-pipeline/pyproject.toml
[project]
name = "agri-rag-pipeline"
version = "0.1.0"
description = "Agriculture RAG Pipeline with LlamaIndex, ChromaDB, and Gemini"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-vertex>=0.1.0",
    # python library connector between LlamaIndex and open-source ChromaDB
    "llama-index-vector-stores-chroma>=0.1.0",
    
    # Vector database
    "chromadb>=0.4.0",
    
    # Google Cloud
    "google-cloud-storage>=2.10.0",
    "google-cloud-aiplatform>=1.38.0",
    "vertexai>=1.38.0",
    
    # PDF processing
    "pdfplumber>=0.10.0",
    
    # Text processing
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    
    # Utilities
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    
    # Text cleaning (for HTML if needed later)
    "trafilatura>=1.6.0",
    "ftfy>=6.1.0",
    "charset-normalizer>=3.3.0",
]

[dependency-groups]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
]